requirements.txt
loguru
pydantic
pydantic-settings
google-cloud
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
pymysql
pyyaml
flask
gunicorn
text_unidecode
pandas